changed components to pages, added side bar menu, and fitted divs to screen
