--- src/components/Products/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Products() {
-  return (
-    <div>
-      <h1>Products Page</h1>
-    </div>
-  );
-}
-
-export default Products;
