--- src/components/Sales/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Sales() {
-  return (
-    <div>
-      <h1>Sales Page</h1>
-    </div>
-  );
-}
-
-export default Sales;
