--- src/components/Stats/index.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Stats() {
-  return (
-    <div>
-      <h1>Stats Page</h1>
-    </div>
-  );
-}
-
-export default Stats;
