+++ src/pages/Landing.js
@@ -2,20 +2,17 @@ import React from "react";
 import "./style.css";
 import Buttons from "../components/Buttons";
 import DataTable from "../components/DataTable";
-import Sales from "../components/Sales";
+import Sales from "./Sales";
 
 function Landing() {
   return (
-      <div>
-        
-        <Buttons />
+    <div>
+      <Buttons />
 
-        <DataTable />
+      <DataTable />
 
-        <Sales />
-
-      
-      </div>
+      <Sales />
+    </div>
   );
 }
