+++ src/pages/Products.js
@@ -1,0 +1,15 @@
+import React from "react";
+import Menu from "../components/Menu";
+
+function Products() {
+  return (
+    <div id="productsPage">
+      <Menu />
+      <div id="productsWrapper">
+        <h1>Products Page</h1>
+      </div>
+    </div>
+  );
+}
+
+export default Products;
