+++ src/pages/Sales.js
@@ -1,0 +1,15 @@
+import React from "react";
+import Menu from "../components/Menu";
+
+function Sales() {
+  return (
+    <div id="salesPage">
+      <Menu />
+      <div id="salesWrapper">
+        <h1>Sales Page</h1>
+      </div>
+    </div>
+  );
+}
+
+export default Sales;
