+++ src/pages/style.css
@@ -8,7 +8,18 @@
 }
  */
 
-#employeePage {
+#employeePage,
+#salesPage,
+#productsPage,
+#statsPage {
   border: solid red 2px;
-  max-width: 88%;
+  width: 88%;
+  float: right;
+}
+
+#empWrapper,
+#salesWrapper,
+#productsWrapper,
+#statsWrapper {
+  border: solid blue 2px;
 }
